Add new class for core data instance
